Use 'requests' instead of 'urllib{,2,3}'
This gets rid of the necessity for most of the 'if PYTHON_VERSION...'
cruft, as well as cleaning up the code around the requests themselves.

I'm certain that there is more tidying to do here, but for now, this
seems to work for all the endpoints I've tried, on both python3.4.3 and
python2.7.9 against my v4 controller.